同步代码

+++ NetworkCore/src/include/Tool.hpp
@@ -4,11 +4,11 @@
 #include <cstdint>
 #include <cstring>
 
-#include "Session.hpp"
+#include "entity/Buffer.hpp"
 namespace Tool {
 uint16_t readRingBufferUint16(Buffer& buffer) {
-    uint16_t value = (static_cast<uint8_t>((buffer.data[buffer.head]) << 8) |
-                      static_cast<uint8_t>(
+    uint16_t value = (static_cast<uint16_t>((buffer.data[buffer.head]) << 8) |
+                      static_cast<uint16_t>(
                           buffer.data[(buffer.head + 1) % buffer.data.size()]));
     buffer.head = (buffer.head + 2) % buffer.data.size();
     buffer.size -= 2;
@@ -16,14 +16,14 @@ uint16_t readRingBufferUint16(Buffer& buffer) {
 }
 
 uint32_t readRingBufferUint32(Buffer& buffer) {
-    uint32_t value = (static_cast<uint8_t>(buffer.data[buffer.head]) << 24) |
-                     (static_cast<uint8_t>(
+    uint32_t value = (static_cast<uint32_t>(buffer.data[buffer.head]) << 24) |
+                     (static_cast<uint32_t>(
                           buffer.data[(buffer.head + 1) % buffer.data.size()])
                       << 16) |
-                     (static_cast<uint8_t>(
+                     (static_cast<uint32_t>(
                           buffer.data[(buffer.head + 2) % buffer.data.size()])
                       << 8) |
-                     static_cast<uint8_t>(
+                     static_cast<uint32_t>(
                          buffer.data[(buffer.head + 3) % buffer.data.size()]);
     buffer.head = (buffer.head + 4) % buffer.data.size();
     buffer.size -= 4;
